first start at typed jira client
