+++ src/util/get-cloud-or-server.ts
@@ -1,4 +1,9 @@
 import { GITHUB_CLOUD_API_BASEURL } from "utils/get-github-client-config";
 
-export const getCloudOrServerFromGitHubAppId = (gitHubAppId: number | undefined) => gitHubAppId ? "server" : "cloud";
-export const getCloudOrServerFromHost = (host: string) => GITHUB_CLOUD_API_BASEURL.includes(host) ? "cloud" : "server";
+export enum GithubProductEnum {
+	CLOUD = "cloud",
+	SERVER = "server",
+}
+
+export const getCloudOrServerFromGitHubAppId = (gitHubAppId: number | undefined): GithubProductEnum => gitHubAppId ? GithubProductEnum.SERVER : GithubProductEnum.CLOUD;
+export const getCloudOrServerFromHost = (host: string): GithubProductEnum => GITHUB_CLOUD_API_BASEURL.includes(host) ? GithubProductEnum.CLOUD : GithubProductEnum.SERVER;
